Safer is iov_len is long.
git-svn-id: http://svn.research.cc.gatech.edu/kaos/ffs/trunk@12992 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ fm/fm_internal.h
@@ -243,7 +243,7 @@ typedef int (*IOinterface_func)(void *conn, void *buffer, int length,
 #define HAVE_IOVEC_DEFINE
 struct	iovec {
     const void *iov_base;
-    int	iov_len;
+    long iov_len;
 };
 #endif
